untested stage to blackbox config types
