+++ server/src/controllers/projects.rs
@@ -1,80 +1,21 @@
-use aws_sdk_s3::{config, types, Client, Credentials, Region};
+use aws_sdk_s3::types;
 
-extern crate r2d2_redis;
-use r2d2_redis::{r2d2, redis, RedisConnectionManager};
-use r2d2_redis::redis::Commands;
-use r2d2::PooledConnection;
-
-use aws_smithy_http;
-use rocket::{Outcome, Request, State, get};
-use rocket::request::{self, FromRequest};
+use rocket::get;
 use rocket_contrib::json::Json;
-use rocket::http::Status;
 
 use ron::de::from_str;
 use ron::ser::{to_string_pretty, PrettyConfig};
 
-use crate::config::db;
+use crate::config::{db, types as db_types};
 use crate::models::{Project, Slides};
-
 use crate::mods::main::{print_type_of, unzip_from_buff};
 
 extern crate dotenv;
 use std::env::var;
-use std::ops::{Deref, DerefMut};
-
-#[derive(Debug)]
-pub enum S3Errors {
-    StreamErr(aws_smithy_http::byte_stream::Error),
-    FutureErr(types::SdkError<aws_sdk_s3::error::GetObjectError>),
-    RedisErr(redis::RedisError)
-}
-
-impl From<redis::RedisError> for S3Errors {
-    fn from(err: redis::RedisError) -> Self {
-        S3Errors::RedisErr(err)
-    }
-}
-
-impl From<aws_smithy_http::byte_stream::Error> for S3Errors {
-    fn from(err: aws_smithy_http::byte_stream::Error) -> Self {
-        S3Errors::StreamErr(err)
-    }
-}
-
-impl From<types::SdkError<aws_sdk_s3::error::GetObjectError>> for S3Errors {
-    fn from(err: types::SdkError<aws_sdk_s3::error::GetObjectError>) -> Self {
-        S3Errors::FutureErr(err)
-    }
-}
-
-pub struct Conn(PooledConnection<RedisConnectionManager>);
-impl<'a, 'r> FromRequest<'a, 'r> for Conn {
-    type Error = ();
-
-    fn from_request(request: &'a Request<'r>) -> request::Outcome<Conn, ()> {
-        let pool = request.guard::<State<r2d2::Pool<RedisConnectionManager>>>()?;
-        match pool.get() {
-            Ok(database) => Outcome::Success(Conn(database)),
-            Err(_) => Outcome::Failure((Status::ServiceUnavailable, ())),
-        }
-    }
-}
-impl Deref for Conn {
-    type Target = PooledConnection<RedisConnectionManager>;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-impl DerefMut for Conn {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
 
 
 #[get("/")]
-pub fn read_all(mut redis_conn: Conn) -> Json<Vec<Project>> {
+pub fn read_all(mut redis_conn: db_types::Conn) -> Json<Vec<Project>> {
     let mut proj_result = Vec::new();
 
     if redis_conn.exists("projects").unwrap() {
